whisper-api/requirements.txt
flask
openai-whisper
# requests
# torch
# torchaudio
